Move start practice test action to tests actions
