+++ src/modules/tests/tests.actions.ts
@@ -6,6 +6,7 @@ export const PERSIST_TESTS = '[Tests] Persist';
 export const LOAD_PERSISTED_TESTS = '[Tests] Load persisted';
 export const LOAD_PERSISTED_TESTS_SUCCESS = '[Tests] Load persisted success';
 export const SET_ACTIVITY_CODE = '[Tests] Set activity code';
+export const START_PRACTICE_TEST = '[Tests] Start practice test';
 
 export class PersistTests implements Action {
   readonly type = PERSIST_TESTS;
@@ -25,8 +26,14 @@ export class SetActivityCode implements Action {
   constructor(public payload: ActivityCode) {}
 }
 
+export class StartPracticeTest implements Action {
+  readonly type = START_PRACTICE_TEST;
+  constructor(public slotId: number) { }
+}
+
 export type Types =
   | PersistTests
   | LoadPersistedTests
   | LoadPersistedTestsSuccess
-  | SetActivityCode;
+  | SetActivityCode
+  | StartPracticeTest;
