+++ src/modules/tests/tests.effects.ts
@@ -5,6 +5,10 @@ import { TestPersistenceProvider } from '../../providers/test-persistence/test-p
 import { from } from 'rxjs/observable/from';
 import * as testActions from './tests.actions';
 import { of } from 'rxjs/observable/of';
+import { TestSlotAttributes, TestCentre, Candidate } from '@dvsa/mes-test-schema/categories/B';
+import { PopulateApplicationReference } from './application-reference/application-reference.actions';
+import { PopulateCandidateDetails } from './candidate/candidate.actions';
+import { Application } from '../../shared/models/DJournal';
 
 @Injectable()
 export class TestsEffects {
@@ -36,4 +40,79 @@ export class TestsEffects {
     )),
   );
 
+  @Effect()
+  startPracticeTestEffect$ = this.actions$.pipe(
+    ofType(testActions.START_PRACTICE_TEST),
+    switchMap(() => {
+      const slot = {
+        slotData: {
+          ...practiceSlot,
+        },
+      };
+
+      return [
+        new PopulateApplicationReference(slot.slotData.booking.application),
+        new PopulateCandidateDetails(slot.slotData.booking.candidate),
+      ];
+    }),
+  );
+
+  extractTestSlotAttributes = (slotData): TestSlotAttributes => {
+    return {
+      welshTest: slotData.booking.application.welshTest,
+      slotId: slotData.slotDetail.slotId,
+      start: slotData.slotDetail.start,
+      specialNeeds: slotData.booking.application.specialNeeds,
+      vehicleSlotType: slotData.vehicleSlotType,
+      extendedTest: slotData.booking.application.extendedTest,
+    };
+  }
+
+  extractTestCentre = (slotData): TestCentre => {
+    return {
+      costCode: slotData.testCentre.costCode,
+    };
+  }
+
 }
+
+export const application: Application = {
+  applicationId: 1234569,
+  bookingSequence: 1,
+  checkDigit: 9,
+  entitlementCheck: false,
+  extendedTest: false,
+  progressiveAccess: false,
+  testCategory: 'B',
+  vehicleGearbox: 'Manual',
+  welshTest: false,
+};
+
+export const candidate: Candidate = {
+  candidateAddress: {
+    addressLine1: 'My House',
+    addressLine2: 'Someplace',
+    addressLine3: 'Sometown',
+    postcode: 'AB45 6CD',
+  },
+  candidateId: 1,
+  candidateName: {
+    firstName: 'Practice',
+    lastName: 'Mode',
+    title: 'Miss',
+  },
+  driverNumber: 'MODEX625220A99HC',
+  mobileTelephone: '07654 123456',
+};
+
+const practiceSlot = {
+  slotDetail: {
+    slotId: 1,
+    duration: 57,
+    start: '2019-01-01T10:14:00+00:00',
+  },
+  booking: {
+    application,
+    candidate,
+  },
+};
