+++ src/modules/tests/tests.reducer.ts
@@ -57,7 +57,7 @@ export function testsReducer(
 const deriveSlotId = (state: TestsModel, action: Action): string | null => {
   if (action instanceof journalActions.StartTest
       || action instanceof journalActions.ActivateTest
-      || action instanceof journalActions.StartPracticeTest) {
+      || action instanceof testActions.StartPracticeTest) {
     return `${action.slotId}`;
   }
   return (state.currentTest && state.currentTest.slotId) ? state.currentTest.slotId : null;
